Fix continent with max death count query
For continent with highest death count ,the query was incorrectly showing the death count for the country which has maximum death in that continent. This PR fixes the query to show the continent with maximum death. Death in each continent is calculated by summing the new deaths for each day.